Somewhat reasonable lib/main seperation
