+++ src/diagnostics.rs
@@ -10,7 +10,7 @@ pub struct FileId(pub usize);
 pub struct RtError(pub Diagnostic<usize>);
 
 impl Display for RtError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, _: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // This error should be reported via codespan, not Display
         unreachable!("THIS SHOULD NEVER COME UP. PLEASE FILE A BUG")
     }
@@ -22,7 +22,7 @@ impl std::error::Error for RtError {}
 pub struct CompError(pub Diagnostic<usize>);
 
 impl Display for CompError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, _: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // This error should be reported via codespan, not Display
         unreachable!("THIS SHOULD NEVER COME UP. PLEASE FILE A BUG")
     }
